Move functions into juju utilities.
